refactor!: split state.particle module
