+++ expertsystem/state/__init__.py
@@ -1,11 +1,13 @@
 """All modules related to a quantum mechanical state."""
 
 __all__ = [
+    "combinatorics",
     "conservation_rules",
     "propagation",
     "properties",
 ]
 
+from . import combinatorics
 from . import conservation_rules
 from . import propagation
 from . import properties
